requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==1.65.2
pydantic==2.10.6
python-dotenv==1.0.1
supabase==2.12.0